Transfering tests from DescriptorNameDescriptionOverrideTests to BenchmarkConverterTests and remaking them in a correct way
